democracy_exe/agentic/studio/requirements.txt
langgraph
langchain-core
langchain-community
langchain-openai
langchain-fireworks
pinecone-client
tiktoken
loguru
typing-extensions
python-dotenv
uuid
aiohttp
discord.py
rich
pydantic_settings
yarl
